ml/food_com/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.9.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2023.4.0
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h5py==3.11.0
idna==3.7
Jinja2==3.1.2
joblib==1.4.0
keras==3.2.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pgvector==0.2.5
pillow==10.3.0
protobuf==4.25.3
psycopg==3.1.18
psycopg-binary==3.1.18
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
SQLAlchemy==2.0.29
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
torch==2.2.2+cpu
torchvision==0.17.2+cpu
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
